requirements.txt
fastapi[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings